Move MQTT implementation and add feature checks
